requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openai==1.3.7
pydantic==2.5.0
python-dotenv==1.0.0
twilio==8.10.3
httpx==0.25.2
numpy==1.25.2
fuzzywuzzy==0.18.0
python-levenshtein==0.21.1